refactored table view cells
