+++ travel-weather/View Controllers/WeatherListViewController.swift
@@ -13,14 +13,19 @@ import UIKit
 
 
 class WeatherListViewController: UITableViewController {
-
+    
+    //MARK:- Properties
+    
     var stateController: StateController! {
         didSet {
             stateController.weatherListDelegate = self
         }
     }
-    
+    //default number of rows if there is no "No Internet" Warning
     var numberOfRows = 15
+    
+    
+    //MARK:- Lifecycle
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -33,17 +38,17 @@ class WeatherListViewController: UITableViewController {
         tableView.clipsToBounds = false
         tableView.backgroundColor = .systemGray6
         
-        var cellNib = UINib(nibName: "DayCell", bundle: nil)
-        tableView.register(cellNib, forCellReuseIdentifier: "DayCell")
-        cellNib = UINib(nibName: "DarkSkyAttributionCell", bundle: nil)
-        tableView.register(cellNib, forCellReuseIdentifier: "AttributionCell")
-        tableView.register(UINib.init(nibName: "NoInternetCell", bundle: nil), forCellReuseIdentifier: "NoInternetCell")
+        tableView.register(UINib(nibName: "DayCell", bundle: nil), forCellReuseIdentifier: "DayCell")
+        tableView.register(UINib(nibName: "DarkSkyAttributionCell", bundle: nil), forCellReuseIdentifier: "AttributionCell")
+        tableView.register(UINib(nibName: "NoInternetCell", bundle: nil), forCellReuseIdentifier: "NoInternetCell")
         tableView.reloadData()
         
         
     }
     
-    func configureCell(day: Day, cell: DayCell) {
+    
+    //MARK:- Internal functions
+    private func configureCell(day: Day, cell: DayCell) {
         cell.weekdayLabel.text = day.weekday
         cell.dateLabel.text = DateHelper.shortDateFormat(date: day.date)
         cell.cityLabel.text = day.location.display
@@ -52,8 +57,6 @@ class WeatherListViewController: UITableViewController {
         cell.weatherImageView.image = day.weatherImage
         cell.colorView.backgroundColor = stateController.getAssociatedColor(for: day.location.placeID)
     }
-        
-    
     
     
     //MARK:- Table View Data Source
@@ -108,6 +111,7 @@ class WeatherListViewController: UITableViewController {
 }
 
 extension WeatherListViewController: StateControllerDelegate {
+    
     func didUpdateForecast() {
         tableView.reloadData()
     }
