+++ travel-weather/Views/DayCell.swift
@@ -11,16 +11,11 @@ import UIKit
 class DayCell: WeatherListCell {
     
     //MARK:- Outlets
-    @IBOutlet weak var weekdayLabel: UILabel!
-    @IBOutlet weak var cityLabel: UILabel!
-    @IBOutlet weak var weatherImageView: UIImageView!
+    @IBOutlet var weekdayLabel: UILabel!
+    @IBOutlet var cityLabel: UILabel!
+    @IBOutlet var weatherImageView: UIImageView!
     @IBOutlet var dateLabel: UILabel!
-    
-    
-    @IBOutlet weak var highTempLabel: UILabel!
-    
-    @IBOutlet weak var lowTempLabel: UILabel!
-    
-
+    @IBOutlet var highTempLabel: UILabel!
+    @IBOutlet var lowTempLabel: UILabel!
     
 }
