+++ travel-weather/Views/WeatherListCell.swift
@@ -15,18 +15,22 @@ class WeatherListCell: UITableViewCell {
     
     override func awakeFromNib() {
         super.awakeFromNib()
+        
         colorView.layer.cornerRadius = 13
         colorView.layer.shadowColor = UIColor.black.cgColor
         colorView.layer.shadowOffset = CGSize(width: 0,  height: 3)
         colorView.layer.shadowOpacity = 0.4
         colorView.layer.shadowRadius = 2
-        contentView.layer.masksToBounds = false
-        contentView.clipsToBounds = false
         colorView.clipsToBounds = false
         colorView.layer.masksToBounds = false
-        self.clipsToBounds = false
-        self.layer.masksToBounds = false
-        backgroundColor = .systemGray6
         colorView.backgroundColor = .charcoalGrayLight
+        
+        contentView.layer.masksToBounds = false
+        contentView.clipsToBounds = false
+        
+        clipsToBounds = false
+        layer.masksToBounds = false
+        backgroundColor = .systemGray6
+        
        }
 }
